+++ semver-gradle-plugin/main/kotlin/com/javiersc/semver/gradle/plugin/internal/git/GitRef.kt
@@ -1,17 +1,17 @@
 package com.javiersc.semver.gradle.plugin.internal.git
 
-internal sealed interface GitRef {
+internal sealed class GitRef {
 
-    data class Head(val commit: Commit) : GitRef
+    data class Head(val commit: Commit) : GitRef()
 
-    data class Commit(val message: String, val fullMessage: String, val hash: String) : GitRef
+    data class Commit(val message: String, val fullMessage: String, val hash: String) : GitRef()
 
-    data class Tag(val name: String, val refName: String, val commit: Commit) : GitRef
+    data class Tag(val name: String, val refName: String, val commit: Commit) : GitRef()
 
     data class Branch(
         val name: String, // example: `main`
         val refName: String, // example: `refs/heads/main`
         val commits: List<Commit>,
         val tags: List<Tag>,
-    ) : GitRef
+    ) : GitRef()
 }